backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "fastapi>=0.119.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "workos>=5.31.2",
]
